refactor(addon-sdk): soft cutover to packages/addon-sdk/ + shim
Per ADR-038 \u00a712.1 C12 (Public SDK External Boundary), the V0.1
add-on SDK source moves to packages/addon-sdk/ and src/sdk/addons/
becomes a re-export shim pointing at the new package. Single source
of truth immediately; consumers don't notice the change.

This is a soft, fork-only cutover. The package is not yet published
to npm; package.json documents the intended public boundary for the
future Phase 1 extraction.

packages/addon-sdk/ added:
  README.md       (114 lines: scope, layout, compatibility, REF vocab,
                   reference implementations, see-also)
  package.json    (@resonantos/addon-sdk 0.1.0, intent field pointing
                   at ADR-038 C12 and ADR-040, exports map for the
                   five entry points)
  src/contracts.ts        (moved from src/sdk/addons/, 85 lines)
  src/validation.ts       (moved from src/sdk/addons/, 1180 lines)
  src/registry.ts         (moved from src/sdk/addons/, 150 lines)
  src/surface-routing.ts  (moved from src/sdk/addons/, 60 lines)
  src/index.ts            (moved from src/sdk/addons/, 33 lines)

Import paths inside the moved files updated from
'../../core/contracts' to '../../../src/core/contracts' so they
resolve correctly from packages/addon-sdk/src/.

src/sdk/addons/* turned into re-export shims:
  contracts.ts       -> packages/addon-sdk/src/contracts
  validation.ts      -> packages/addon-sdk/src/validation
  registry.ts        -> packages/addon-sdk/src/registry
  surface-routing.ts -> packages/addon-sdk/src/surface-routing
  index.ts           -> packages/addon-sdk/src/index

src/sdk/addons/*.test.ts unchanged: tests still import from ./contracts,
./validation, ./registry, ./surface-routing (now shims), exercising the
new package transparently.

tsconfig.json include extended to ['src', 'packages/addon-sdk/src',
'vite.config.ts'].

docs:
- docs/architecture/ADR-038-resonant-extension-framework.md \u00a712 C12
  row records the cutover as 'landed (soft, fork-only)' and links to
  ADR-040 for the external-agent-runtime boundary.
- docs/addons/resonant-extension-framework/README.md files section
  pointer to packages/addon-sdk/README.md.

Not in this commit (deferred to follow-on):
- packages/addon-sdk-testing/ (mock host + add-on test fixtures)
- npm workspaces declaration
- src/sdk/addons/ shim removal
- Phase 1 actual workspace extraction (publishable package, install
  wiring, CI install/audit blocks)

Verified:
- npx tsc --noEmit: clean
- npm run docs:check: passes
- npx vitest run: 429/429 tests pass (shims exercise new package)
