+++ packages/addon-sdk/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@resonantos/addon-sdk",
+  "version": "0.1.0",
+  "description": "Public Resonant Extension Framework (REF) add-on SDK \u2014 manifest contract, capability model, validator, registry, and surface-routing helpers. See ADR-038 \u00a712.1 C12 (Public SDK External Boundary) and ADR-040.",
+  "license": "MIT",
+  "type": "module",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts",
+    "./contracts": "./src/contracts.ts",
+    "./validation": "./src/validation.ts",
+    "./registry": "./src/registry.ts",
+    "./surface-routing": "./src/surface-routing.ts"
+  },
+  "files": [
+    "src/"
+  ],
+  "engines": {
+    "node": ">=22.13.0"
+  },
+  "keywords": [
+    "resonantos",
+    "add-on",
+    "sdk",
+    "ref",
+    "manifest",
+    "validation"
+  ],
+  "intent": "Fork-only soft cutover per ADR-038 \u00a712.1 C12. The package is not yet published; this manifest records the intended public boundary. When the SDK is extracted to a real workspace (Phase 1 of the implementation roadmap), the source files in src/ become the published surface and the existing src/sdk/addons/ shims in this repo are removed.",
+  "sourceLocation": "src/sdk/addons/ (shims \u2014 keep for backward compat); canonical: packages/addon-sdk/src/",
+  "seeAlso": [
+    "docs/architecture/ADR-038-resonant-extension-framework.md",
+    "docs/architecture/ADR-040-provider-fabric-boundary-external-agent-runtimes.md",
+    "docs/addons/resonant-extension-framework/RESONANT_ADDON_SDK_SPEC_V0.1.md",
+    "docs/addons/resonant-extension-framework/ADDON_PACKAGE_AND_MANIFEST_SPEC_V0.1.md"
+  ]
+}
