+++ src/sdk/addons/contracts.ts
@@ -1,85 +1,7 @@
-// Intent citation: docs/architecture/ADR-018-addon-sdk-v0.md
+// Re-export shim \u2014 public SDK surface lives at packages/addon-sdk/.
+// This shim keeps existing internal imports working during the soft cutover.
+// See ADR-038 \u00a712.1 C12 row "Public SDK External Boundary" and ADR-040 \u00a73.
+//
+// Source of truth: packages/addon-sdk/src/contracts.ts
 
-import type {
-  AddOnAugmentorSkill,
-  AddOnEngineerSetupRunbook,
-  AddOnLocalServiceDefinition,
-  AddOnManifest,
-  AddOnAgentRuntimeContract,
-  AddOnAuditContract,
-  AddOnConnectorDefinition,
-  AddOnEmbeddedWorkspaceContract,
-  AddOnHookDefinition,
-  AddOnInstallContract,
-  AddOnMemoryAccessContract,
-  AddOnScriptDefinition,
-  AddOnServiceProtocol,
-  AddOnSkillDefinition,
-  AddOnDeterministicSmokeTest,
-  AddOnToolDefinition,
-  AddOnWorkflowBoundary,
-  Capability,
-} from "../../core/contracts";
-
-export const ADDON_SDK_VERSION = "0.1.0";
-
-export type AddOnSdkManifest = AddOnManifest & {
-  sdkVersion: string;
-  service?: AddOnLocalServiceDefinition;
-  tools?: AddOnToolDefinition[];
-  workflowBoundaries?: AddOnWorkflowBoundary[];
-  skills?: AddOnSkillDefinition[];
-  connectors?: AddOnConnectorDefinition[];
-  scripts?: AddOnScriptDefinition[];
-  hooks?: AddOnHookDefinition[];
-  engineerSetup?: AddOnEngineerSetupRunbook;
-  augmentorSkills?: AddOnAugmentorSkill[];
-  install?: AddOnInstallContract;
-  audit?: AddOnAuditContract;
-  embeddedWorkspace?: AddOnEmbeddedWorkspaceContract;
-  agentRuntime?: AddOnAgentRuntimeContract;
-  memoryAccess?: AddOnMemoryAccessContract;
-  smokeTests?: AddOnDeterministicSmokeTest[];
-};
-
-export type AddOnManifestSource = "bundled" | "sideload";
-
-export type AddOnValidationSeverity = "error" | "warning";
-
-export type AddOnValidationIssue = {
-  severity: AddOnValidationSeverity;
-  code: string;
-  path: string;
-  message: string;
-};
-
-export type AddOnManifestValidationResult = {
-  valid: boolean;
-  manifestId?: string;
-  issues: AddOnValidationIssue[];
-};
-
-export const ADDON_CAPABILITIES: readonly Capability[] = [
-  "filesystem",
-  "archive-read",
-  "archive-intake-write",
-  "chat-interface",
-  "memory-provider",
-  "providers",
-  "shell",
-  "network",
-  "ui-embedding",
-  "browser-control",
-  "agent-delegation",
-  "notifications",
-  "device-integration",
-  "channel.send",
-  "channel.account-write",
-];
-
-export const ADDON_SERVICE_PROTOCOLS: readonly AddOnServiceProtocol[] = [
-  "stdio-json-rpc",
-  "http-json",
-  "websocket-json",
-  "host-command",
-];
+export * from "../../../packages/addon-sdk/src/contracts";
