+++ src/sdk/addons/registry.ts
@@ -1,149 +1,7 @@
-// Intent citation: docs/architecture/ADR-018-addon-sdk-v0.md
-// Intent citation: docs/architecture/ADR-023-addon-repository-registry-model.md
+// Re-export shim \u2014 public SDK surface lives at packages/addon-sdk/.
+// This shim keeps existing internal imports working during the soft cutover.
+// See ADR-038 \u00a712.1 C12 row "Public SDK External Boundary" and ADR-040 \u00a73.
+//
+// Source of truth: packages/addon-sdk/src/registry.ts
 
-import type {
-  AddOnArtifactReference,
-  AddOnInstallation,
-  AddOnManifest,
-  AddOnProvenanceTier,
-  AddOnRegistryEntry,
-  AddOnRegistryReviewState,
-  AddOnRegistrySource,
-  InstallationStatus,
-  ManifestVerificationState,
-} from "../../core/contracts";
-
-export interface AddOnRegistryEntryOptions {
-  registrySource: AddOnRegistrySource;
-  installation?: AddOnInstallation;
-  manifestRef?: Partial<AddOnArtifactReference>;
-  releaseArtifact?: AddOnArtifactReference;
-  sourceRepositoryUrl?: string;
-  reviewState?: AddOnRegistryReviewState;
-  notes?: string[];
-}
-
-export interface AddOnRegistryBuildInput {
-  bundled: AddOnManifest[];
-  sideloaded: AddOnManifest[];
-  installations: Record<string, AddOnInstallation>;
-}
-
-export interface AddOnRegistrySnapshot {
-  entries: AddOnRegistryEntry[];
-  byId: Record<string, AddOnRegistryEntry>;
-}
-
-const sourceDefaults = (
-  manifest: AddOnManifest,
-  registrySource: AddOnRegistrySource,
-): {
-  provenanceTier: AddOnProvenanceTier;
-  verificationState: ManifestVerificationState;
-  reviewState: AddOnRegistryReviewState;
-} => {
-  if (registrySource === "sideloaded-local" || registrySource === "developer-local") {
-    return {
-      provenanceTier: "sideloaded-unverified",
-      verificationState: "unverified",
-      reviewState: "unreviewed",
-    };
-  }
-
-  if (!manifest.provenance) {
-    return {
-      provenanceTier: "sideloaded-unverified",
-      verificationState: "unverified",
-      reviewState: "unreviewed",
-    };
-  }
-
-  return {
-    // Trust is never granted by omission (#345): a PRESENT but partial provenance
-    // block falls back to the least-trusted value for each missing field, and is
-    // only "reviewed" when both tier and verificationState are explicit. An ABSENT
-    // block is handled above (P1-e).
-    provenanceTier: manifest.provenance.tier ?? "sideloaded-unverified",
-    verificationState: manifest.provenance.verificationState ?? "unverified",
-    reviewState: manifest.provenance.tier && manifest.provenance.verificationState ? "reviewed" : "unreviewed",
-  };
-};
-
-const defaultManifestRef = (manifest: AddOnManifest, registrySource: AddOnRegistrySource): AddOnArtifactReference => ({
-  type: "manifest",
-  label: `${manifest.name} manifest`,
-  path: registrySource === "bundled-catalog" ? "/addons/index.json" : undefined,
-  signatureRef: manifest.provenance?.signatureRef,
-});
-
-const installStateFromInstallation = (installation: AddOnInstallation | undefined): {
-  installState: InstallationStatus;
-  installed: boolean;
-  enabled: boolean;
-} => ({
-  installState: installation?.status ?? "available",
-  installed: installation?.installed ?? false,
-  enabled: installation?.enabled ?? false,
-});
-
-export const createAddOnRegistryEntry = (
-  manifest: AddOnManifest,
-  options: AddOnRegistryEntryOptions,
-): AddOnRegistryEntry => {
-  const defaults = sourceDefaults(manifest, options.registrySource);
-  const installState = installStateFromInstallation(options.installation);
-  const manifestRef = {
-    ...defaultManifestRef(manifest, options.registrySource),
-    ...options.manifestRef,
-    type: "manifest" as const,
-  };
-
-  return {
-    addonId: manifest.id,
-    name: manifest.name,
-    version: manifest.version,
-    author: manifest.author,
-    category: manifest.category,
-    description: manifest.description,
-    runtimeType: manifest.runtimeType,
-    registrySource: options.registrySource,
-    provenanceTier: options.installation?.provenanceTier ?? defaults.provenanceTier,
-    verificationState: options.installation?.verificationState ?? defaults.verificationState,
-    reviewState: options.reviewState ?? defaults.reviewState,
-    manifestRef,
-    releaseArtifact: options.releaseArtifact,
-    sourceRepositoryUrl: options.sourceRepositoryUrl,
-    compatibility: manifest.compatibility,
-    requestedCapabilities: manifest.requestedCapabilities.map((capability) => ({ ...capability })),
-    recommendedGrantPresetIds:
-      options.installation?.recommendedGrantPresetIds ?? (manifest.grantPresets ?? []).map((preset) => preset.id),
-    ...installState,
-    notes: options.notes ?? options.installation?.notes ?? ["Catalog entry is not installed yet."],
-  };
-};
-
-export const createAddOnRegistrySnapshot = ({
-  bundled,
-  sideloaded,
-  installations,
-}: AddOnRegistryBuildInput): AddOnRegistrySnapshot => {
-  const entries = [
-    ...bundled.map((manifest) =>
-      createAddOnRegistryEntry(manifest, {
-        registrySource: "bundled-catalog",
-        installation: installations[manifest.id],
-      }),
-    ),
-    ...sideloaded.map((manifest) =>
-      createAddOnRegistryEntry(manifest, {
-        registrySource: "sideloaded-local",
-        installation: installations[manifest.id],
-      }),
-    ),
-  ];
-
-  return {
-    entries,
-    byId: Object.fromEntries(entries.map((entry) => [entry.addonId, entry])),
-  };
-};
+export * from "../../../packages/addon-sdk/src/registry";
