+++ tsconfig.json
@@ -17,5 +17,5 @@
     "jsx": "react-jsx",
     "types": ["vitest/globals"]
   },
-  "include": ["src", "vite.config.ts"]
+  "include": ["src", "packages/addon-sdk/src", "vite.config.ts"]
 }
